1.  Fixed naming collision of 'clock' variables by renaming them to 'mx_clock'.
2.  Updated Debian 5 support on PowerPC for PowerPMAC emulation testing.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@5101 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_record.c
@@ -2519,7 +2519,7 @@ mx_print_all_field_attributes( FILE *output, MX_RECORD *record )
 		}
 
 		fprintf(output, "%7ld ", field->num_elements);
-		fprintf(output, "%7ld ", field->max_bytes);
+		fprintf(output, "%7ld ", (long) field->max_bytes);
 
 		fprintf(output, "\n");
 	}
